Prevent user from removing the last element in a JsonArray in RemoveArrayElementCommand.cs.
We are doing this because the AddArrayElementCommand (which will be introduced in a later commit) relies on having atleast one element in the array (this is a bit jank, more on that in a later commit). In addition, having no elements in an array will cause the serializer to go nuts.